09-user-feedback/requirements.txt
openai
python-dotenv
arize-phoenix
arize-phoenix-client
arize-phoenix-evals
elastic-opentelemetry
opentelemetry-instrumentation-httpx
openinference-instrumentation-openai
